+++ python/ray/tests/BUILD
@@ -318,6 +318,7 @@ py_test_module_list(
     "test_runtime_env_working_dir.py",
     "test_runtime_env_working_dir_2.py",
     "test_runtime_env_working_dir_3.py",
+    "test_runtime_env_working_dir_4.py",
     "test_runtime_env_working_dir_remote_uri.py"
   ],
   size = "large",
